+++ findmseqseed/findmseqseed.cpp
@@ -1,7 +1,7 @@
 #include "mex.h"
 #include "matrix.h"
 #include "matlabaux.h"
-
+#include <inttypes.h>
 #ifdef _WIN32
 #include <windows.h>
 
@@ -26,30 +26,30 @@ void mexFunction(int nlhs, mxArray *plhs[],
 	}
 	else
 	{
-		unsigned long nbits;
+		uint32_t nbits;
 		mwSize dim[2] = { 1, 1};
 
-		nbits = maGetUINT32Element(prhs[0], 0); //*((unsigned long*)mxGetData(prhs[0]));
+		nbits = maGetUINT32Element(prhs[0], 0); //*((uint32_t*)mxGetData(prhs[0]));
 		dim[0] = (1<<nbits);
 
 
 		int cnt = 0;
 
 		printf("searching for m-sequence seeds...\n");
-		unsigned long nbitstat = nbits - 4;
+		uint32_t nbitstat = nbits - 4;
 	
 		// # of seeds is less than half the possible choices (for sure, it's actually a lot less)
 		// this saves some time because we won't have to dynamically allocate.
 		// there is probably a formula to calculate the exact # of good seeds.
-		unsigned long* seeds = new unsigned long[(1<<(nbits-1))];
+		uint32_t* seeds = new uint32_t[(1<<(nbits-1))];
 
 		// this is just an exhaustive search
 		for (int j = 1; j < (1<<nbits); j=j+2)
 		{
-			unsigned long reg1 = 1;
-			unsigned long reg2 = j;
-			unsigned long np1bitmask = dim[0];
-			unsigned long outputmask = dim[0] - 1;
+			uint32_t reg1 = 1;
+			uint32_t reg2 = j;
+			uint32_t np1bitmask = dim[0];
+			uint32_t outputmask = dim[0] - 1;
 			int i;
 
 			// using sutter's tap register method (when reg1 repeats, the sequence repeats)
@@ -79,8 +79,8 @@ void mexFunction(int nlhs, mxArray *plhs[],
 		// allocate matlab array to pass data back and copy data to array
 		dim[0] = cnt;
 		plhs[0] = mxCreateNumericArray(sizeof(dim) / sizeof(dim[0]), dim, mxUINT32_CLASS, mxREAL);
-		unsigned long* result = (unsigned long*)mxGetData(plhs[0]);
-		memcpy(result, seeds, cnt * sizeof(unsigned long));
+		uint32_t* result = (uint32_t*)mxGetData(plhs[0]);
+		memcpy(result, seeds, cnt * sizeof(uint32_t));
 
 		// finally free up temporary memory
 		delete [] seeds;
