+++ gridprewalshperm/gridprewalshperm.cpp
@@ -41,15 +41,15 @@ void mexFunction(int nlhs, mxArray *plhs[],
 		int count = maGetUINT32Element(prhs[2], 0);
 		mwSize dim[2] = { 1<<order, 1};
 		int i;
-		unsigned long reg1 = 1;
-		unsigned long reg2;
-		unsigned long datasize = mxGetM(prhs[3]) * mxGetN(prhs[3]);
+		uint32_t reg1 = 1;
+		uint32_t reg2;
+		uint32_t datasize = mxGetM(prhs[3]) * mxGetN(prhs[3]);
 		int elementsize;
 		char* result;
 		char* indata;
 		elementsize = datasize / dim[0];
 #ifdef ERROR_CHECK
-		if (datasize != (unsigned long)(dim[0]*elementsize))
+		if (datasize != (uint32_t)(dim[0]*elementsize))
 		{
 			mexErrMsgTxt("invalid parameter.  data needs to have k*2^order elements for integer k.");
 			return;
@@ -64,14 +64,14 @@ void mexFunction(int nlhs, mxArray *plhs[],
 		result = (char*)mxGetData(plhs[0]);
 		indata = (char*)mxGetData(prhs[2]);
 
-		unsigned long* newbit = NULL;
+		uint32_t* newbit = NULL;
 
 		// if look up table is available in parameter, then use it
 		if (nrhs >= 5)
-			newbit = (unsigned long*)mxGetData(prhs[4]);
+			newbit = (uint32_t*)mxGetData(prhs[4]);
 		else // if not available, then create it
 		{
-			newbit = new unsigned long[dim[0]];
+			newbit = new uint32_t[dim[0]];
 			for (i = 0; i < dim[0]; i++)
 			{
 				int j;	
@@ -87,13 +87,13 @@ void mexFunction(int nlhs, mxArray *plhs[],
 		// then do the pre walsh transform permutation as
 		// specified in Sutter's SIAM paper.
 		reg1 = 1;
-		unsigned long outputmask = dim[0] - 1;
+		uint32_t outputmask = dim[0] - 1;
 
 		// the first element stays the same
 		memcpy(result, indata, elementsize);
 		indata += elementsize;
 
-		unsigned long* temp = new unsigned long[dim[0]];
+		uint32_t* temp = new uint32_t[dim[0]];
 
 		// for the rest, use the sequences in the generating register
 		// scheme to create the permutation!
